Reorganize files; change main layout
