+++ app/src/main/java/com/example/cattlelog/main_activity/HerdListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.cattlelog.herd
+package com.example.cattlelog.main_activity
 
 import android.content.Context
 import android.view.LayoutInflater
